move tests of BIDS command utilities into test_utility module
